% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col_widths.R
\name{add_col_widths}
\alias{add_col_widths}
\title{Adds width as a attribute to each column from a list of tibbles}
\usage{
add_col_widths(tbl_list, fwf_pos)
}
\arguments{
\item{tbl_list}{a list of tibbles}

\item{fwf_pos}{a list of returns from \code{fwf_position()} with
length that matches length of \code{tbl_list}}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Adds width as a attribute to each column from a list of tibbles
}
\keyword{internal}
