% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSAT_tbl_df_class.R
\name{as_DSSAT_tbl}
\alias{as_DSSAT_tbl}
\title{Convert tibble to DSSAT_tbl}
\usage{
as_DSSAT_tbl(tbl_in, v_fmt = NULL, tier_info = NULL)
}
\arguments{
\item{tbl_in}{a tibble}

\item{v_fmt}{a character vector specifying the sprintf() format for each column}

\item{tier_info}{a list of character vectors storing the history of which original
table that columns came from for tibbles that are comprised of multiple joined tables}
}
\value{
a tibble of class DSSAT_tbl
}
\description{
Convert tibble to DSSAT_tbl
}
