% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TAV_AMP.R
\name{calc_AMP}
\alias{calc_AMP}
\title{Calculate long-term temperature amplitude (AMP)}
\usage{
calc_AMP(wth)
}
\arguments{
\item{wth}{a data frame that contains weather data formatted
as would be generated using \link{read_wth}, namely
including columns DATE, TMAX and TMIN.}
}
\description{
Calculate long-term temperature amplitude (AMP)
}
