% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TAV_AMP.R
\name{calc_TAV}
\alias{calc_TAV}
\title{Calculate long-term temperature average (TAV)}
\usage{
calc_TAV(wth)
}
\arguments{
\item{wth}{a data frame that contains weather data formatted
as would be generated using \link{read_wth}, namely
including columns TMAX and TMIN.}
}
\description{
Calculate long-term temperature average (TAV)
}
