% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_output.R
\name{clear_output}
\alias{clear_output}
\title{Clear output files}
\usage{
clear_output(dir_name = getwd(), file_ext = c("OUT", "LST", "INP", "INH"))
}
\arguments{
\item{dir_name}{a length-one character vector indicating the directory in
which to delete output files (by default the current working directory)}

\item{file_ext}{a character vector of file extensions to delete}
}
\description{
A function to delete DSSAT output files (\*.OUT) and intermediate files
(\*.INH, \*.INP and \*.LST) from a directory
}
