% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tiers.R
\name{combine_tiers}
\alias{combine_tiers}
\title{Efficiently combined data from two data tiers into one}
\usage{
combine_tiers(tier1, tier2)
}
\arguments{
\item{tier1, tier2}{a tibble for each tier of data}
}
\value{
a tibble containing the data from the combined tiers
}
\description{
Efficiently combined data from two data tiers into one
}
\keyword{internal}

