% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_date.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Converts a column to POSIXct}
\usage{
convert_to_date(d)
}
\arguments{
\item{d}{a numeric or character vector}
}
\value{
a POSIXct vector
}
\description{
Converts a column to POSIXct
}
\examples{

convert_to_date('2018031')

}
\keyword{internal}
