% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cul_template.R
\name{cul_template}
\alias{cul_template}
\title{Generate template cultivar parameter data frame}
\usage{
cul_template(model_code, title = NULL, ...)
}
\arguments{
\item{model_code}{a character value that contains the DSSAT-CSM model code}

\item{title}{a character value of the first line header for the cultivar
file (optional)}
}
\description{
Generate template cultivar parameter data frame
}
