% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSSAT-package.R
\docType{package}
\name{DSSAT}
\alias{DSSAT}
\title{DSSAT: A Comprehensive R Interface for the DSSAT Cropping Systems Model}
\description{
The DSSAT package provides a comprehensive
R interface to the Decision Support System for Agrotechnology
Transfer Cropping Systems Model (DSSAT-CSM; see
<https://dssat.net> for more information).
The package provides cross-platform functions to read and
write input files, run DSSAT-CSM, and read output files.
}
\section{DSSAT functions}{

The mypackage functions ...
}

