% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_template.R
\name{eco_template}
\alias{eco_template}
\title{Generate template ecotype parameter data frame}
\usage{
eco_template(model_code, title = NULL, ...)
}
\arguments{
\item{model_code}{a character value that contains the DSSAT-CSM model code}

\item{title}{a character value of the first line header for the cultivar
file (optional)}
}
\description{
Generate template ecotype parameter data frame
}
