% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template.R
\name{filex_template}
\alias{filex_template}
\title{Generate File X template}
\usage{
filex_template(...)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate File X template
}
