% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_cultivars.R
\name{filex_template_cultivars}
\alias{filex_template_cultivars}
\title{Generate a template CULTIVARS section of File X}
\usage{
filex_template_cultivars(
  CR = "FA",
  INGENO = "IB0001",
  CNAME = "",
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate a template CULTIVARS section of File X
}
