% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_fertilizers.R
\name{filex_template_fertilizers}
\alias{filex_template_fertilizers}
\title{Generate a template FERTILIZERS section of File X}
\usage{
filex_template_fertilizers(
  F = NA_real_,
  FDATE = as.POSIXct(NA),
  FMCD = NA_character_,
  FACD = NA_character_,
  FDEP = NA_real_,
  FAMN = NA_real_,
  FAMP = NA_real_,
  FAMK = NA_real_,
  FAMC = NA_real_,
  FAMO = NA_real_,
  FOCD = NA_character_,
  FERNAME = NA_character_,
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to use [tidyr::expand_grid()]
to generate combinations of any input values supplied as a vector}
}
\description{
Generate a template FERTILIZERS section of File X
}
