% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_fields.R
\name{filex_template_fields}
\alias{filex_template_fields}
\title{Generate a template FEILDS section of File X}
\usage{
filex_template_fields(
  ID_FIELD = NA_character_,
  WSTA = NA_character_,
  FLSA = NA_character_,
  FLOB = NA_real_,
  FLDT = NA_character_,
  FLDD = NA_real_,
  FLDS = NA_real_,
  FLST = NA_character_,
  SLTX = NA_character_,
  SLDP = NA_real_,
  ID_SOIL = NA_character_,
  FLNAME = NA_character_,
  XCRD = NA_real_,
  YCRD = NA_real_,
  ELEV = NA_real_,
  AREA = NA_real_,
  SLEN = NA_real_,
  FLWR = NA_real_,
  SLAS = NA_real_,
  FLHST = NA_character_,
  FHDUR = NA_real_,
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate a template FEILDS section of File X
}
