% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_general.R
\name{filex_template_general}
\alias{filex_template_general}
\title{Generate a template GENERAL section of File X}
\usage{
filex_template_general(
  PEOPLE = NA_character_,
  ADDRESS = NA_character_,
  SITE = NA_character_,
  PAREA = NA_real_,
  PRNO = NA_real_,
  PLEN = NA_real_,
  PLDR = NA_real_,
  PLSP = NA_real_,
  PLAY = NA_character_,
  HAREA = NA_real_,
  HRNO = NA_real_,
  HLEN = NA_real_,
  HARM = NA_character_,
  NOTES = NA_character_
)
}
\description{
Generate a template GENERAL section of File X
}
