% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_planting_details.R
\name{filex_template_planting_details}
\alias{filex_template_planting_details}
\title{Generate a template PLANTING DETAILS section of File X}
\usage{
filex_template_planting_details(
  PDATE = as.POSIXct(NA),
  EDATE = as.POSIXct(NA),
  PPOP = NA_real_,
  PPOE = NA_real_,
  PLME = NA_character_,
  PLDS = NA_character_,
  PLRS = NA_real_,
  PLRD = NA_real_,
  PLDP = NA_real_,
  PLWT = NA_real_,
  PAGE = NA_real_,
  PENV = NA_real_,
  PLPH = NA_real_,
  SPRL = NA_real_,
  PLNAME = NA_character_,
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate a template PLANTING DETAILS section of File X
}
