% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_simulation_controls.R
\name{filex_template_simulation_controls}
\alias{filex_template_simulation_controls}
\title{Generate a template SIMULATION CONTROLS section of File X}
\usage{
filex_template_simulation_controls(
  GENERAL = "GE",
  NYERS = 1,
  NREPS = 1,
  START = "S",
  SDATE = NA_character_,
  RSEED = 2150,
  SNAME = NA_character_,
  SMODEL = NA_character_,
  OPTIONS = "OP",
  WATER = "N",
  NITRO = "N",
  SYMBI = "N",
  PHOSP = "N",
  POTAS = "N",
  DISES = "N",
  CHEM = "N",
  TILL = "N",
  CO2 = "M",
  METHODS = "ME",
  WTHER = "M",
  INCON = "M",
  LIGHT = "E",
  EVAPO = "R",
  INFIL = "S",
  PHOTO = "L",
  HYDRO = "R",
  NSWIT = 1,
  MESOM = "G",
  MESEV = "R",
  MESOL = 2,
  MANAGEMENT = "MA",
  PLANT = "R",
  IRRIG = "N",
  FERTI = "N",
  RESID = "N",
  HARVS = "M",
  OUTPUTS = "OU",
  FNAME = "N",
  OVVEW = "Y",
  SUMRY = "Y",
  FROPT = 1,
  GROUT = "N",
  CAOUT = "N",
  WAOUT = "N",
  NIOUT = "N",
  MIOUT = "N",
  DIOUT = "N",
  VBOSE = "N",
  CHOUT = "N",
  OPOUT = "N",
  FMOPT = "A",
  PLANTING = "PL",
  PFRST = NA_character_,
  PLAST = NA_character_,
  PH2OL = NA_real_,
  PH2OU = NA_real_,
  PH2OD = NA_real_,
  PSTMX = NA_real_,
  PSTMN = NA_real_,
  IRRIGATION = "IR",
  IMDEP = NA_real_,
  ITHRL = NA_real_,
  ITHRU = NA_real_,
  IROFF = NA_character_,
  IMETH = NA_character_,
  IRAMT = NA_real_,
  IREFF = NA_real_,
  NITROGEN = "NI",
  NMDEP = NA_real_,
  NMTHR = NA_real_,
  NAMNT = NA_real_,
  NCODE = NA_character_,
  NAOFF = NA_character_,
  RESIDUES = "RE",
  RIPCN = NA_real_,
  RTIME = NA_real_,
  RIDEP = NA_real_,
  HARVEST = "HA",
  HFRST = NA_character_,
  HLAST = NA_character_,
  HPCNP = NA_real_,
  HPCNR = NA_real_,
  SIMDATES = "SI",
  ENDAT = NA_character_,
  SDUR = NA_real_,
  FODAT = NA_character_,
  FSTRYR = NA_real_,
  FENDYR = NA_real_,
  FWFILE = NA_character_,
  FONAME = NA_character_,
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate a template SIMULATION CONTROLS section of File X
}
