% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filex_template_treatments.R
\name{filex_template_treatments}
\alias{filex_template_treatments}
\title{Generate a template TREATMENTS section of File X}
\usage{
filex_template_treatments(
  R = 0,
  O = 0,
  C = 0,
  TNAME = "",
  CU = 1,
  FL = 1,
  SA = 0,
  IC = 0,
  MP = 1,
  MI = 0,
  MF = 0,
  MR = 0,
  MC = 0,
  MT = 0,
  ME = 0,
  MH = 0,
  SM = 1,
  expand = FALSE
)
}
\arguments{
\item{expand}{a logical value indicating whether or not to generate expanded
combinations of any input values supplied as a vector}
}
\description{
Generate a template TREATMENTS section of File X
}
