% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_to_fwf_position.R
\name{header_to_fwf_position}
\alias{header_to_fwf_position}
\title{Converts header line into a vector of column names}
\usage{
header_to_fwf_position(header)
}
\arguments{
\item{header}{A length-one character vector of a header line
from a DSSAT format output file}
}
\value{
a tibble containing the begin position, end position,
 and name for each column from the original header line
}
\description{
Converts header line into a vector of column names
}
\examples{

header_line <- '@YEAR DOY   DAS   VAR1  DAYLD   TWLD   SRAD'
header_to_fwf_position(header_line)
}

