% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_justified.R
\name{left_justified_cul}
\alias{left_justified_cul}
\title{Return list of left justified columns in DSSAT CUL files}
\usage{
left_justified_cul()
}
\description{
Return list of left justified columns in DSSAT CUL files
}
