% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_justified.R
\name{left_justified_filex}
\alias{left_justified_filex}
\title{Return list of left justified columns in DSSAT FileX}
\usage{
left_justified_filex()
}
\description{
Return list of left justified columns in DSSAT FileX
}
