% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_justified.R
\name{left_justified_soil}
\alias{left_justified_soil}
\title{Return list of left justified columns in DSSAT Soil file}
\usage{
left_justified_soil()
}
\description{
Return list of left justified columns in DSSAT Soil file
}
