% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sol.R
\name{link_soil_comments}
\alias{link_soil_comments}
\title{Helper function to link comments to pedon}
\usage{
link_soil_comments(comments_lines, pedon_start_end)
}
\description{
Helper function to link comments to pedon
}
