% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_cond.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Convenience function that allows mutating a subset of rows}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{a tibble}

\item{condition}{a logical vector for subsetting rows of `.data`}

\item{...}{Name-value pairs of expressions to be evaluated by
`mutate()`}

\item{envir}{environment within which expressions should be evaluated}
}
\value{
a tibble with specified rows modified
}
\description{
Convenience function that allows mutating a subset of rows
}
\details{
Original code taken from \url{https://stackoverflow.com/questions/34096162/dplyr-mutate-replace-several-columns-on-a-subset-of-rows}
}
