% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_to_regex.R
\name{name_to_regex}
\alias{name_to_regex}
\title{Convert column names to a regex}
\usage{
name_to_regex(cnames)
}
\arguments{
\item{cnames}{A character vector that contains column names}
}
\value{
A character vector with escaped special characters (e.g.
parentheses, periods, asterisks)
}
\description{
Convert column names to a regex
}
\examples{

name_to_regex("TB(1)")

}
\keyword{internal}
