% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phenology_section.R
\name{parse_cropgro_phenology_section}
\alias{parse_cropgro_phenology_section}
\title{Parses phenology section of a DSSAT CROPGRO species file (*.SPE)}
\usage{
parse_cropgro_phenology_section(lines)
}
\arguments{
\item{lines}{a character vector}
}
\value{
a list containing the data from the raw DSSAT file
}
\description{
Parses phenology section of a DSSAT CROPGRO species file (*.SPE)
}
\keyword{internal}

