% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cropgro_spe_line.R
\name{parse_cropgro_spe_line}
\alias{parse_cropgro_spe_line}
\title{Reads data from one line of a DSSAT CROPGRO species file (*.SPE)}
\usage{
parse_cropgro_spe_line(line, sec_name = "")
}
\arguments{
\item{line}{a length-one character vector}

\item{sec_name}{a length-one character vector}
}
\value{
a list containing the data from the raw line
}
\description{
Reads data from one line of a DSSAT CROPGRO species file (*.SPE)
}
\keyword{internal}

