% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_leaf_senescence_section.R
\name{parse_leaf_senescence_section}
\alias{parse_leaf_senescence_section}
\title{Parses leaf senescence section of a DSSAT CROPGRO species file (*.SPE)}
\usage{
parse_leaf_senescence_section(lines)
}
\arguments{
\item{lines}{a character vector}
}
\value{
a list containing the data from the raw DSSAT file
}
\description{
Parses leaf senescence section of a DSSAT CROPGRO species file (*.SPE)
}
\keyword{internal}

