% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phosphorus_section.R
\name{parse_phosphorus_section}
\alias{parse_phosphorus_section}
\title{Parses phosphorus section of a DSSAT species file (*.SPE)}
\usage{
parse_phosphorus_section(lines)
}
\arguments{
\item{lines}{a character vector}
}
\value{
a list containing the data from the raw DSSAT file
}
\description{
Parses phosphorus section of a DSSAT species file (*.SPE)
}
\keyword{internal}

