% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_casupro_eco.R
\name{read_casupro_eco}
\alias{read_casupro_eco}
\title{Reads parameters from a single DSSAT CASUPRO ecotype parameter file (*.ECO)}
\usage{
read_casupro_eco(file_name)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads parameters from a single DSSAT CASUPRO ecotype parameter file (*.ECO)
}
\examples{

sample_eco <- c('CA0001  (CP 80-1743; CP 88-1762)',
'!-------------------',
'  KCAN_ECO      0.80    Superseeds KCAN from .SPE file',
'  SMAX            20    Maximum number of stalks a variety can yield per stubble or plant (limit 41)',
'!-------------------',
'  CAB              8    Number of leaves in spindle cluster at top of stalk ("cabbage").  In CASUPRO, stalk dry matter begins to accumulate only after CAB leaves have emerged. ',
'  CanLmtFac    0.010    Proportion of stalk leaf area moved up each day to keep de-populated light interception zones full.',
'  LfShdFac      0.35    Proportion of a stalks new leaf area that shades (displaces to the next lower zone) leaves on other stalks.',
'  MINSHD        0.15    Minimum value allowed for FRSHAV (Average of daylight fraction of ground area in plant shadow) during early growth.(0.01-0.20)',
'!-------------------',
'  StkSenFrac  .00005    Threshold maintenance respiration deficit to stalk dry weight ratio (TotalDEF(Stalk) / STKWT(DAS-1,Stalk) at which stalk death occurs. ',
'  SENDAY        0.05    Maximum fraction of existing leaf weight which can be senesced on a single day as a function of severe water stress 4 days earlier.  ')

write(sample_eco,'SAMPLE.ECO')

read_casupro_eco('SAMPLE.ECO')
}
\keyword{internal}

