% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cropgro_spe.R
\name{read_cropgro_spe}
\alias{read_cropgro_spe}
\title{Parses a DSSAT CROPGRO species file (*.SPE)}
\usage{
read_cropgro_spe(filename)
}
\arguments{
\item{filename}{a length-one character vector}
}
\value{
a list containing the data from the raw DSSAT file


(?!\\![^\\*])'
}
\description{
Parses a DSSAT CROPGRO species file (*.SPE)
}
\keyword{internal}

