% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filet.R
\name{read_filet}
\alias{read_filet}
\title{Reads time series data from a single DSSAT file T}
\usage{
read_filet(file_name, col_types = NULL, col_names = NULL,
  na_strings = NULL)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}

\item{na_strings}{a character vector of string to represent missing values}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads time series data from a single DSSAT file T
}
\examples{

sample_filet <- c(
"*EXP.DATA (T): SATO8901 SC LAMERCY RAINFED SUGARCANE VALIDATION EXP.      ",
"",
"@TRNO  DATE  LGDMD LAITD SUCMD SMDMD  L#SD  T#AD",
"     1 90032 22.40  3.00  2.10  8.40  8.00  20.0",
"     1 90093 35.00  4.30  7.60 19.40 10.10  16.6",
"     1 90155 47.00  3.40 14.10 30.20 11.30  15.0",
"     1 90213 52.10  2.70 17.10 33.80  6.50  15.1")

\dontrun{

write(sample_filet,'SAMPLE.CRT')

read_filet('SAMPLE.CRT')

}
}
