% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filex.R
\name{read_filex}
\alias{read_filex}
\title{Reads input data from a single DSSAT experiment file (*.*X)}
\usage{
read_filex(
  file_name,
  col_types = NULL,
  col_names = NULL,
  na_strings = NULL,
  store_v_fmt = FALSE
)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}

\item{na_strings}{A character vector containing strings that should be interpreted as
missing values}

\item{store_v_fmt}{a logical value indicating whether or not to store the
format for variables being read}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads input data from a single DSSAT experiment file (*.*X)
}
