% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filex_multiple.R
\name{read_filex_multiple}
\alias{read_filex_multiple}
\title{Read multiple File X}
\usage{
read_filex_multiple(
  file_name,
  col_types = NULL,
  col_names = NULL,
  na_strings = NULL,
  store_v_fmt = FALSE,
  use_std_fmt = TRUE
)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}

\item{na_strings}{A character vector containing strings that should be interpreted as
missing values}

\item{store_v_fmt}{a logical value indicating whether or not to store the
format for variables being read}

\item{use_std_fmt}{logical value indicating whether to read
FileX using the standard column formats. If FALSE, column
formats will be inferred from tier headers}
}
\description{
Read multiple File X
}
