% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pest.R
\name{read_pest}
\alias{read_pest}
\title{Reads input data from a single DSSAT pest file (*.PST)}
\usage{
read_pest(file_name, col_types = NULL, col_names = NULL)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads input data from a single DSSAT pest file (*.PST)
}
