% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sol.R
\name{read_sol}
\alias{read_sol}
\title{Reads soil parameters from a single DSSAT soil parameter file (*.SOL)}
\usage{
read_sol(file_name, id_soil = NULL, nested = TRUE)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{id_soil}{a length-one character vector containing the soil ID code for a
single soil profile}

\item{nested}{a logical to set whether layer data should be nested in the output
(one row per soil profile)}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads soil parameters from a single DSSAT soil parameter file (*.SOL)
}
\examples{

# Extract file path for sample soil file
sample_sol <- system.file('extdata','SAMPLE.SOL',package='DSSAT')

sol <- read_sol(sample_sol)

}
