% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wth.R
\name{read_wth}
\alias{read_wth}
\title{Reads weather input data from a single DSSAT weather file (*.WTH)}
\usage{
read_wth(file_name, col_types = NULL, col_names = NULL)
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads weather input data from a single DSSAT weather file (*.WTH)
}
\examples{

sample_wth <- c(
"*WEATHER DATA : Gainesville,Florida,USA",
"",
"@ INSI      LAT     LONG  ELEV   TAV   AMP REFHT WNDHT",
"  UFGA   29.630  -82.370    10  21.3  13.0   2.0   3.0",
"@DATE  SRAD  TMAX  TMIN  RAIN  DEWP  WIND   PAR  EVAP  RHUM",
"06001   7.6  26.8  14.1   0.5                          ",
"06002   9.5  27.8  19.4   0.5                          ",
"06003   6.1  24.9  11.7   0.0                          ",
"06004  12.5  21.8   6.6   0.0                          ",
"06005  12.3  22.3   7.5   0.0                          ")

write(sample_wth,'EXAMPLE.WTH')

read_wth('EXAMPLE.WTH')
}
