% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cul.R
\name{write_cul}
\alias{write_cul}
\title{Reads parameters from a single DSSAT cultivar parameter file (*.CUL)}
\usage{
write_cul(cul, file_name)
}
\arguments{
\item{cul}{a DSSAT_tbl containing the contents of a DSSAT cultivar parameter file}

\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}
}
\value{
a tibble containing the data from the raw DSSAT output
}
\description{
Reads parameters from a single DSSAT cultivar parameter file (*.CUL)
}
\examples{

sample_cul <- c(
"*MAIZE CULTIVAR COEFFICIENTS: MZCER047 MODEL",
"!",
"@VAR#  VRNAME.......... EXPNO   ECO#    P1    P2    P5    G2    G3 PHINT",
"!                                        1     2     3     4     5     6",
"PC0001 2500-2600 GDD        . IB0001 160.0 0.750 780.0 750.0  8.50 49.00",
"PC0002 2600-2650 GDD        . IB0001 185.0 0.750 850.0 800.0  8.50 49.00",
"PC0003 2650-2700 GDD        . IB0001 212.0 0.750 850.0 800.0  8.50 49.00")

\dontrun{

write(sample_cul,'SAMPLE.CUL')

cul <- read_cul('SAMPLE.CUL')

write_cul(cul,'SAMPLE2.CUL')

}
}
