% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dssbatch.R
\name{write_dssbatch}
\alias{write_dssbatch}
\title{Constructs and writes a DSSAT simulation batch file}
\usage{
write_dssbatch(batch_tbl = NULL, filex = "", trtno = 1, rp = 1,
  sq = 0, op = 0, co = 0, file_name = NULL)
}
\arguments{
\item{batch_tbl}{a data frame or tibble that contains all
required columns of a DSSAT batch file (FILEX, TRTNO, RP,
SQ, OP, CO)}

\item{filex}{a character vector}

\item{trtno, rp, sq, op, co}{a numeric vector}

\item{file_name}{an optional character vector of the intended
batch file name}
}
\value{
invisibly returns a character vector containing the
content of a DSSAT batch file
}
\description{
Constructs and writes a DSSAT simulation batch file
}
\examples{

write_dssbatch(filex='UFGA0601.BMX',trtno=1:4)

batch_tbl <- tibble(FILEX='UFGA0601.BMX',
                    TRTNO=1:4,
                    RP=1,
                    SQ=0,
                    OP=0,
                    CO=0)

write_dssbatch(batch_tbl)

}
