% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dssbatch.R
\name{write_dssbatch.tbl_df}
\alias{write_dssbatch.tbl_df}
\title{tbl_df method for write_dssbatch()}
\usage{
\method{write_dssbatch}{tbl_df}(x, trtno = 1, rp = 1, sq = 0, op = 0, co = 0, file_name = NULL)
}
\arguments{
\item{x}{a tibble/data frame or character vector; if a tibble, it should contain all
required columns of a DSSAT batch file (FILEX, TRTNO, RP, SQ, OP, CO);
if a character vector, it should contain FileX file names}

\item{trtno, rp, sq, op, co}{a numeric vector}

\item{file_name}{an optional character vector of the intended
batch file name}
}
\description{
tbl_df method for write_dssbatch()
}
\keyword{internal}
