% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tier.R
\name{write_tier}
\alias{write_tier}
\title{Writes data from a single DSSAT data tier}
\usage{
write_tier(tier_data)
}
\arguments{
\item{tier_data}{a tibble containing the data to write out}
}
\value{
a character vector
}
\description{
Writes data from a single DSSAT data tier
}
\examples{

tier_data <- tibble(TRNO=1:4,HWAM=rnorm(4,2000,250))

tier_data <- add_v_fmt(tier_data,v_fmt=c(TRNO='\%6.0f', HWAM='\%6.0f'))

write_tier(tier_data)
}

