% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tiers.R
\name{combine_tiers}
\alias{combine_tiers}
\title{Efficiently combine data from two data tiers into one}
\usage{
combine_tiers(tier1, tier2, use_collapse_rows = FALSE, force_bind_rows = FALSE)
}
\arguments{
\item{tier1, tier2}{a tibble for each tier of data}

\item{use_collapse_rows}{a length-one logical vector that
indicates whether rows should be collapsed before join}
}
\value{
a tibble containing the data from the combined tiers
}
\description{
Efficiently combine data from two data tiers into one
}
\keyword{internal}
