% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eco.R
\name{read_eco}
\alias{read_eco}
\title{Reads parameters from a single DSSAT ecotype parameter file (*.ECO)}
\usage{
read_eco(file_name, col_types = NULL, col_names = NULL,
  left_justified = c("ECO   ", "ECO#", "ECONAME\\\\.*", "ECO-NAME\\\\.*"))
}
\arguments{
\item{file_name}{a character vector of length one that contains the name
of a single DSSAT output file}

\item{col_types}{One of NULL, a cols() specification, or a string.
See \code{\link[readr]{read_fwf}} or \code{vignette("readr")} for
more details.}

\item{col_names}{A character vector of column names; primarily
helpful for cases where there are no white spaces between column
names within the header line}

\item{left_justified}{A character vector of column names that
should be left justified}
}
\value{
a tibble containing the data from the raw DSSAT file
}
\description{
Reads parameters from a single DSSAT ecotype parameter file (*.ECO)
}
\examples{

sample_eco <- c(
"*MAIZE ECOTYPE COEFFICIENTS: MZCER047 MODEL",
"@ECO#  ECONAME.........  TBASE  TOPT ROPT   P20  DJTI  GDDE  DSGFT  RUE   KCAN  TSEN  CDAY",
"!                            1     2    3     4     5     6     7     8      9    10    11",
"IB0001 GENERIC MIDWEST1    8.0 34.0  34.0  12.5   4.0   6.0   170.  4.2   0.85",
"IB0002 GENERIC MIDWEST2    8.0 34.0  34.0  12.5   4.0   6.0   170.  4.5   0.85",
"IB0003 GENERIC MIDWEST3    8.0 34.0  34.0  12.5   4.0   6.0   170.  2.0   0.85",
"IB0004 +5\% RUE MIDWEST1    8.0 34.0  34.0  12.5   4.0   6.0   170.  4.4   0.85",
"DFAULT DEFAULT             8.0 34.0  34.0  12.5   4.0   6.0   170.  4.2   0.85")

write(sample_eco,'SAMPLE.ECO')

read_eco('SAMPLE.ECO')
}

