% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wth.R
\name{write_wth}
\alias{write_wth}
\title{Writes data to a single DSSAT weather file}
\usage{
write_wth(wth, file_name)
}
\arguments{
\item{wth}{a tibble containing the data to write to a DSSAT
weather file}

\item{file_name}{a character vector of length one that contains the name
of a single DSSAT file into which `wth` will be written}
}
\description{
Writes data to a single DSSAT weather file
}
